Moved DS files to designsystem package
