+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/common/composeui/designsystem/WooPosSizes.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.woopos.common.composeui
+package com.woocommerce.android.ui.woopos.common.composeui.designsystem
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
